pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "xrepomirror"
dynamic = ["version"]
description = "A CLI tool to mirror helm charts and docker images from public registries to local/private repositories."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.6"
keywords = ["helm", "docker", "mirror", "registry", "devops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "PyYAML>=5.1",
    "requests>=2.20.0",
]

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
version_scheme = "post-release"
local_scheme = "no-local-version"

[project.scripts]
xrepomirror = "xrepomirror.__main__:main"

[project.urls]
Homepage = "https://github.com/tkdpython/xrepomirror"
Repository = "https://github.com/tkdpython/xrepomirror"
Issues = "https://github.com/tkdpython/xrepomirror/issues"
